Move table aux parsers to its own module
